Dont with GLS part two

+++ src/main/java/app/DaoS/LocationDAO.java
@@ -1,13 +1,15 @@
 package app.DaoS;
 
+import app.DTOMapper.DTOMapper;
 import app.entities.Location;
+import app.records.LocationDTO;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityManagerFactory;
 import jakarta.persistence.TypedQuery;
 
 import java.util.List;
 
-public class LocationDAO implements Idao<Location, Integer> {
+public class LocationDAO implements Idao<Location, Integer> , IDTODAO<LocationDTO, Integer> {
 
     private final EntityManagerFactory emf;
 
@@ -17,6 +19,29 @@ public class LocationDAO implements Idao<Location, Integer> {
 
 
     @Override
+    public LocationDTO getDtoById(Integer id) {
+        try (EntityManager em = emf.createEntityManager()) {
+            Location location = em.find(Location.class, id);
+            return DTOMapper.toLocationDTO(location);
+        } catch (Exception ex) {
+            throw new RuntimeException("Error getting LocationDTO by ID", ex);
+        }
+    }
+
+    @Override
+    public List<LocationDTO> getAllDtos() {
+        try (EntityManager em = emf.createEntityManager()) {
+            TypedQuery<Location> query = em.createQuery("SELECT l FROM Location l", Location.class);
+            List<Location> locations = query.getResultList();
+            return locations.stream()
+                    .map(DTOMapper::toLocationDTO)
+                    .toList();
+        } catch (Exception ex) {
+            throw new RuntimeException("Error getting all LocationDTOs", ex);
+        }
+    }
+
+    @Override
     public Location create(Location location) {
         try (EntityManager em = emf.createEntityManager()) {
             em.getTransaction().begin();
@@ -24,7 +49,7 @@ public class LocationDAO implements Idao<Location, Integer> {
             em.getTransaction().commit();
             return location;
         } catch (Exception ex) {
-            throw new RuntimeException(" Error creatin new location", ex);
+            throw new RuntimeException("Error creating new location", ex);
         }
     }
 
@@ -33,19 +58,20 @@ public class LocationDAO implements Idao<Location, Integer> {
         try (EntityManager em = emf.createEntityManager()) {
             return em.find(Location.class, id);
         } catch (Exception ex) {
-            throw new RuntimeException("Error getting location id ", ex);
+            throw new RuntimeException("Error getting location by ID", ex);
         }
+
     }
 
     @Override
     public Location update(Location location) {
         try (EntityManager em = emf.createEntityManager()) {
             em.getTransaction().begin();
-            em.merge(location);
+            Location updated = em.merge(location);
             em.getTransaction().commit();
-            return location;
+            return updated;
         } catch (Exception ex) {
-            throw new RuntimeException("Error updating location id ", ex);
+            throw new RuntimeException("Error updating location", ex);
         }
     }
 
@@ -54,21 +80,24 @@ public class LocationDAO implements Idao<Location, Integer> {
         try (EntityManager em = emf.createEntityManager()) {
             em.getTransaction().begin();
             Location location = em.find(Location.class, id);
-            em.remove(location);
-            em.getTransaction().commit();
-            return true;
+            if (location != null) {
+                em.remove(location);
+                em.getTransaction().commit();
+                return true;
+            }
+            em.getTransaction().rollback();
+            return false;
         } catch (Exception ex) {
-            throw new RuntimeException("Error deleting location id ", ex);
+            throw new RuntimeException("Error deleting Location", ex);
         }
     }
 
     @Override
     public List<Location> getAll() {
         try (EntityManager em = emf.createEntityManager()) {
-            TypedQuery<Location> query = em.createQuery("select l from Location l", Location.class);
-            return query.getResultList();
+            return em.createQuery("SELECT l FROM Location l", Location.class).getResultList();
         } catch (Exception ex) {
-            throw new RuntimeException(" Error getting all locations", ex);
+            throw new RuntimeException("Error getting all Locations", ex);
         }
     }
 }
